Refactored code to use constants for default particles and populations
Added two new constants, DEFAULT_PARTICLE and DEFAULT_POP, to the library for maintaining a default set of particles and their characteristics. This not only improves readability, but also performance by eliminating the need to repeatedly call clone on particles or use mutable variables. This change enhances the efficiency of our Newtonian physics engine implementation.